Rewrite with new architecture
The old architecture was getting hard to manage. It had some big and
ugly modules, awkward functions, and even wrong formatting for the
error messages.

This commits is part 1 of a bigger rewrite. Currently errors are not
handled, but the modules have been broken down and some APIs are much
cleaner.